sports-participation-service/requirements.txt
fastapi
uvicorn[standard]
pymongo
pyjwt
httpx
motor
pydantic-settings
structlog
redis
